back/requirements.txt
asgiref==3.2.3
Flask==1.1.2
Flask-CORS==3.0.8
Flask-GraphQL==2.0.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
grpcio==1.28.1
gunicorn==20.0.4
httplib2==0.18.1
Jinja2==2.11.2
mypy==0.781
Pyrebase==3.0.27
python-decouple==3.3
python-firebase==1.2
SQLAlchemy==1.3.18
urllib3==1.25.8
Werkzeug==1.0.1
